fix: Exclude vtarget_pwm from ESP8266 builds
- Only compile vtarget_pwm.c for ESP32/ESP32-C3/ESP32-S3 targets
- Fix format string compatibility (%u instead of %lu for uint32_t)
- Resolves ESP8266 build errors with LEDC_AUTO_CLK and ledc_get_duty

+++ main/vtarget_pwm.c
@@ -146,8 +146,8 @@ esp_err_t vtarget_set_voltage(uint16_t voltage_mv)
         return ret;
     }
 
-    ESP_LOGI(TAG, "VTarget set to %d mV (duty cycle: %lu / %d)",
-             voltage_mv, duty_cycle, VTARGET_PWM_MAX_DUTY);
+    ESP_LOGI(TAG, "VTarget set to %d mV (duty cycle: %u / %d)",
+             voltage_mv, (unsigned int)duty_cycle, VTARGET_PWM_MAX_DUTY);
 
     return ESP_OK;
 }
